requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
appdirs==1.4.3
argh==0.26.2
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
bleach==3.1.4
bump2version==0.5.11
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==4.5.4
distlib==0.3.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pathtools==0.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==4.6.5
pytest-runner==5.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinxcontrib-websupport==1.2.1
SQLAlchemy==1.3.16
toml==0.10.0
tqdm==4.45.0
twine==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==20.0.18
watchdog==0.9.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
psycopg2-binary==2.8.5
